refactor: expose terminal runner port
